solartech/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filebrowser==4.0.3
django-grappelli==3.0.8
django-tinymce-4==0.0.27
pillow==10.2.0
python-pip==1.1.1
six==1.16.0
sqlparse==0.4.4
style==1.1.0
typing_extensions==4.10.0
update==0.0.1
